requirements.txt
distlib
dj-database-url
Django
django-captcha
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
environ
pillow
psycopg2-binary
whitenoise
